+++ unimall-app-api/src/main/java/com/iotechn/unimall/app/api/integral/IntegralServiceImpl.java
@@ -74,7 +74,7 @@ public class IntegralServiceImpl implements IntegralService {
         if (!CollectionUtils.isEmpty(recommendAd)) {
             recommendAd = recommendAd.stream().filter(item -> {
                 try {
-                    if (item.getUnionType() == AdvertUnionType.PRODUCT.getCode()) {
+                    if (item.getUnionType().intValue() == AdvertUnionType.PRODUCT.getCode()) {
                         SpuDTO spuDTO = productBizService.getProductByIdFromCache(Long.parseLong(item.getUnionValue()));
                         item.setData(spuDTO);
                         return true;
